closing connection on any read timeout
since we cant guarantee that the other end wont write the response after
we got a read timeout. If that happens then we would get that response
as the response for the next request. This is not a problem for the
sentinel client since the read timeout is only used during
initialization.